refactor: 서비스에서 스터디 생성 시 ID와 참여 코드를 한번에 반환하도록 변경
