--- backend/src/main/java/harustudy/backend/service/dto/CreatePomodoroStudyResponse.java
@@ -1,10 +1,0 @@
-package harustudy.backend.service.dto;
-
-import harustudy.backend.entity.ParticipantCode;
-
-public record CreatePomodoroStudyResponse(String participantCode) {
-
-    public static CreatePomodoroStudyResponse of(ParticipantCode participantCode) {
-        return new CreatePomodoroStudyResponse(participantCode.getCode());
-    }
-}
